add folders controller & schema
