+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const getAll = require("./getAll");
+const getById = require("./getById");
+const add = require("./add");
+const updateById = require("./updateById");
+const deleteContact = require("./delete");
+
+module.exports = {
+    getAll,
+    getById,
+    add,
+    updateById,
+    deleteContact,
+}
