+++ routes/api/contacts.js
@@ -1,117 +1,17 @@
 const express = require("express");
 
-const createError = require("http-errors");
-
-const Joi = require("joi");
-
-const contactOperation = require("../../models/contacts");
-
-const schema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
+const ctrl = require("../../controllers/contacts");
 
 const router = express.Router();
 
-router.get("/", async (req, res, next) => {
-  try {
-    const contacts = await contactOperation.listContacts();
-    res.json({
-      status: "success",
-      code: 200,
-      data: {
-        result: contacts,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/", ctrl.getAll);
 
-router.get("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const contact = await contactOperation.getContactById(id);
-    if (!contact) {
-      throw createError(404, `Contact with ID: ${id} not found`);
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      data: {
-        contact,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:id", ctrl.getById);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = schema.validate(req.body);
-    if (error) {
-      error.status = 400;
-      error.message = "missing required name field";
-      throw error;
-    }
-    const result = await contactOperation.addContact(req.body);
-    res.status(201).json({
-      status: "success",
-      code: 201,
-      data: {
-        result,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", ctrl.add);
 
-router.delete("/:id", async (req, res, next) => {
-  try {
-    const { id } = req.params;
-    const result = await contactOperation.removeContact(id);
-    if (!result) {
-      throw createError(404, `Contact with ID: ${id} not found`);
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      message: "contact deleted",
-      data: {
-        result,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:id", ctrl.deleteContact);
 
-router.put("/:id", async (req, res, next) => {
-  try {
-    const { error } = schema.validate(req.body);
-    if (error) {
-      error.status = 400;
-      error.message = "missing field";
-      throw error;
-    }
-    const { id } = req.params;
-    const result = await contactOperation.updateContact(id, req.body);
-    if (!result) {
-      throw createError(404, `Contact with ID: ${id} not found`);
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      data: {
-        result,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:id", ctrl.updateById);
 
 module.exports = router;
